feat(swap): refactor swap functionality and remove deprecated components
- Renamed `HomeView` to `MarketplaceHomeView` in the market page.
- Introduced a new `SwapView` component for the swap functionality.
- Removed the deprecated `swap-v2` components and related paths.
- Updated navigation paths to reflect the new swap structure.
- Cleaned up imports and improved code readability in the liquidity modal.
