+++ infra/rooch-portal-v2/src/app/trade/market/page.tsx
@@ -1,9 +1,9 @@
-import HomeView from 'src/sections/trade/market/home-view';
+import MarketplaceHomeView from 'src/sections/trade/market/home-view';
 
 export const metadata = {
   title: 'Marketplace | List',
 };
 
 export default function Page() {
-  return <HomeView />;
+  return <MarketplaceHomeView />;
 }
